header

+++ keranjang.php
@@ -10,13 +10,16 @@
     </thead>
     <tbody>
         <?php
+        if (isset($_SESSION['cart']) && count($_SESSION['cart']) > 0) {
         foreach (@$_SESSION['cart'] as $key_produk => $val_produk): ?>
             <tr>
                 <td><?=($key_produk+1)?></td><td><?=$val_produk['nama_buku']?></td><td><?=$val_produk['qty']?></td><td><a href="hapus_cart.php?id=<?=$key_produk?>" class="btn btn-danger"><strong>X</strong></a></td>
             </tr>
-
-
-        <?php endforeach ?>
+        <?php endforeach; 
+        } else{
+            echo "Tidak ada buku di dalam keranjang";
+        }
+         ?>
     </tbody>
 </table>
 <a href="checkout.php" class="btn btn-primary">Check Out</a>
